topsym.sh: multiply the counts by log(length)
One problem that I noticed is the lack of sensible tie-breaking at the
beginning of a list.  I.e. if you do "toplist.sh |tail -65536", you tap
into the middle of a long list with e.g. ncount=129:

    129 _ZNSt8_Rb_treeINSt7__cxx1112basic...
    129 _ZNSt8_Rb_treeISsSt4pairIKSsSsESt10...
    129 _ZNSt8_Rb_treeIiSt4pairIKilESt10...
    129 _ZSt14__copy_move_a2ILb0EPNSt7...
 => 129 _ZSt4copyIPNSt7__cxx1112basic...
    129 _ZTI11KListWidget
    129 _ZTI11QSpacerItem
    129 _ZTI13QOpenGLWidget

Instead of breaking the ties lexicographically, it would make more sense
to take the longer symbols (because the remaining symbols are shorter,
and would take e.g. less hash cycles in the program).  Furthermore,
I believe it is a good idea to apply this kind of weighing for all
symbols throughout the entire list, not just those with ncount=129.
Therefore, I multiply ncount by a logarithmic factor of the length:

    ncount *= log(ShortLen + length)

where ShortLen=16 is a constant for fine-tuning the global tie-breaking /
limiting its effect on short symbols.