topsym.sh: symbols in subpackages contribute sqrt(n)
If a symbol is referenced a few times in a package (in different files),
it counts as one.  This is consistent with the upcoming data format
which stores the unique list of symbols per package.  Furthermore,
we want to downvote the symbols which get high counts due to many
subpackages originating from a single srpm.  To achieve this, we run
rpmelfsym.pl two times and sum unique counts per subpackage and per srpm
(in a sense averaging the two).  I noticed though that this is not good
enough.  For example, you can see many _TRACE_QEMU* symbols on the list,
originating from qemu-* subpackages, 150 of them!

Now, why do we want to downvote the symbols?  They generate the same
hashing costs, etc, but they are package-specific and they may go away
without notice (or renamed into something else).  So it's wrong to
conclude that _TRACE_QEMU* are widely used symbols just because qemu
has been sliced into too many subpackages.  Tomorrow they may vanish
into thin air, and the slot they took remains wasted.  As the Bible
says, fire came down from heaven and consumed the captain of fifty
with his fifty!

So when we see such a symbol whose high count is due to many
subpackages, we should be more willing to consider the alternatives
whose raw counts are almost as good but srpm counts are much better.
To fit something like this into the existing scheme, this commit does
the following:

- Symbols are counted once per package, as previously.
- We then count how many subpackages reference a symbol, per srpm.
  However, we limit its contribution to the total count: only sqrt(n),
  per srpm, are summed up.