topsym.sh: boosted undefined symbols, per srpm
If a symbol is undefined in all subpackages, it must be resolved to
another package, and should obtain more weight in the list of widely
used symbols.  This is to address the problem of _Z1?gen_* and other
gcc symbols.  Note that (2**0.63)*1.28==1.98, so two subpackages still
do not pass, and the total number of symbols in the output should remain
the same.