downvote caml.*_\d+$ symbols
Ocaml seems to append unique numbers to some symbols (based on an
internal counter within each module), and these number aren't stable -
they are likely to change as new definitions are added to the module.

middle_end/variable.ml:
    97  let unique_name t =
    98    t.name ^ "_" ^ (Int.to_string t.name_stamp)

I still don't have a deep understanding of how and why the symbols are
mangled, and how references are matched back to the definitions, but
that's largely unnecessary.  Because the only thing we can do at this
stage is to downvote the mangled symbols.

I noticed that symbols with a big number, e.g. _\d\d\d\d, are more
likely to go away than symbols with small numbers such as _\d\d.
The number of digits of course alludes to the logarithm, so I wind up
with a fancy formula which uses log(the exact value) rather than just
the number of digits.

	if (match(sym, /^caml.*_([0-9]+)$/, A))
	    N *= 99 / (99 + log(99 + A[1]))