try 128k list with 64k C++ quota
I've got a provisional implementation which, for each package,
compresses the list of symbol names in that package.  For now it's
just a shell script and a few small C programs which handle pfor16
compression and front coding.  However, I thoroughly check that I can
actually decode the compressed data to the original input - which takes
a larger part of the script. :-)  So now we can evaluate various
compression schemes and various changes to topsym.list.

Each list of symbols is split into two parts: .sym0 is the intersection
with topsym.list, and .sym1 is the rest of symbols.

$ head 0ad.sym0
FreeUPNPUrls
PR_CreateThread
PR_DestroyCondVar
PR_DestroyLock
PR_GetCurrentThread
PR_JoinThread
PR_Lock
PR_MicrosecondsToInterval
PR_MillisecondsToInterval
PR_NewCondVar
$ head 0ad.sym1
Atlas_DisplayError
Atlas_GLSetCurrent
Atlas_GLSwapBuffers
Atlas_NotifyEndOfFrame
Atlas_SetConfigDirectory
Atlas_SetDataDirectory
Atlas_SetMessagePasser
Atlas_StartWindow
JSID_EMPTY
JSID_EMPTYHANDLE

The .sym0 part is compressed with pfor16 (after the symbols are
numbered, etc.).  The .sym1 part is compressed in two stages: 1) front
coding, much like in locatedb(5), followed by 2) lz4 compression with
a pre-trained dictionary.

With 64k list:

$ cat *.sym0 |wc -lc
4095185 122811562
$ cat *.sym1 |wc -lc
14039986 761561160
$ wc -c *.sym0.pfor16 |tail -1
3393630 total
$ wc -c *.sym1.fz |tail -1
79799093 total

With 128k list:

$ cat *.sym0 |wc -lc
4942733 154337064
$ cat *.sym1 |wc -lc
13192438 730035658
$ wc -c *.sym0.pfor16 |tail -1
4007324 total
$ wc -c *.sym1.fz |tail -1
74799185 total

The benefit is modest: the size of *.sym1.fz is reduced by a factor of
0.937, and the total number of *.sym1 symbols by 0.939.  pfor16 encoding
is slightly more efficient though, at about 6.5 bits per integer (there
is probably no real improvement, the "improvement" is due to a dumb enc63
implementation).  Well, I think the overall benefit is far from negigible:
apart from compression, we have 6% less symbols to hash, and 6% less
table lookups.  The thing is, for a symbol on the list to be useful,
the bar is low: even if it's used very few times, it's already a win.
Because handling a number is SO MUCH easier than handling a string.

What bothers me is that many of the new symbols on the list should not
have been exported in the first place.  For example, qemu and gcc
binaries export a lot of symbols for no good reason.  The only reason
I can think of is a poor man's debugging facility.  A much better way
to debug a binary is to install the corresponding *-debuginfo package.
Also, packages in Fedora come with minisymtab sections which provide
limited debugging information (which may be enough for backtracing in
gdb).  Furthermore, symbols in binaries (unlike in shared libraries)
cannot be interposed: LD_PRELOAD=foo.so won't redefine foo().  So they
are not even true ELF symbols in some respects, all the less reason to
export them.

But I've already devised a clever scheme of handling a list of 128k
uint16_t indexes.  My first observation is that the number of C++
symbols is slightly more than 50%.  Two factors boost C++ symbols:
1) multiplying by log(ShortLen + length(sym)); 2) there are more C++
packages in Sisyphus than in Fedora, because Sisyphus has multiple KDE
versions.  So let's say that in a 128k list, a fair share of C++ symbols
is exactly 50%.

Then we can split the table into two parts.  Packages that are not related
to C++ will have only the first part.  Both parts can be encoded as a
single pfor16 blob, with a single enc63 tail: <n0, n1, pfor16 blob>.
But n0 and n1 need to be encoded separately.  When C++ symbols are
present, the function to process uint16_t indexes will be called twice,
like this:

    const struct { ... } builtin_sym[128<<10];

    unsigned n0, n1;
    uint16_t *tab = decode_tab(src, &n0, n1);

    if (n0)
	process(tab, n0, builtin_sym);
    if (n1)
	process(tab + n0, n1, builtin_sym + (64<<10));

So let's try 128k I thought, we can always go back to 64k.

I also limit the number of caml.*_\d+$ symbols to 4k.