Revert "regen with Tumbleweed/x86_64"
Tumbleweed is not a full repo.  For one, it lacks 0ad (this was the
first thing I noticed).  It also lacks firefox.  There are a number
of "additional" repositories and "community repositories", including
Packman which seems to override some part of Tumbleweed proper.
It isn't clear which repos I should use to construct a unified repo
like Fedora.

As a result, we get wrong symbols weights.  For example, a few MPI_*
symbols get a boost.  But this is only because some other common symbols
are referenced mostly in additional repos.